requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
fastapi
uvicorn
python-multipart
google-cloud-speech
google-auth
google-cloud-core
python-multipart
itsdangerous
sqlalchemy
google-cloud-firestore
python-alipay-sdk
wechatpy
qrcode
pillow
redis
